requirements.txt
PyYAML
wikidataintegrator
pandas
tqdm
more_itertools